Move the authorizer in its own file
